==> notebooks/4.1-kv-subject2-pipeline.ipynb <==
# Classifying Drug Administration on Subject 2 (Ach-AT)

## Loading and Processing Data
Using the functions written in `src/data/load_data.py`, we convert the raw signals into a format containing discrete time windows.  
This format is suitable for tsfresh to then be used to extract relevant features.

from src.data.load_data import *

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt


from tsfresh import extract_features
from tsfresh.feature_extraction import MinimalFCParameters

ach_at_02 = load_MEA_data(folder = "data/raw/Ach-AT-subject-02")
# take middle horizontal strip of electrodes
label_MEA_data(ach_at_02, output='ach_at_subject_02')

df = pd.read_hdf("data/processed/ach_at_subject_02.h5")

y = (df[['window_id','y']]
     .drop_duplicates()
     .set_index('window_id')
     .T
     .squeeze()
     .sort_index(0))

cols = ['t', 'window_id', 15, 14, 13,  4, 57, 48, 47, 46]
df_middle = df[cols]
df = df.drop(columns = ['y'])

X = extract_features(df, column_id='window_id', column_sort='t', default_fc_parameters=MinimalFCParameters())
X.to_hdf('achat_02_min.h5', key = 'features', complevel = 9)

X_middle = extract_features(df_middle, column_id='window_id', column_sort='t', default_fc_parameters=MinimalFCParameters())
X_middle.to_hdf('achat_02_min_middle.h5', key = 'features', complevel = 9)

## Model Fitting

Compare the performance of decision trees (individual, interpretable) and random forests (ensemble).
Fit both types of model on each set of features and identify which combination results in the highest classification accuracy.

from sklearn import tree
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import cross_val_score

from tsfresh import select_features
from tsfresh.utilities.dataframe_functions import impute

### Feature set 1: Middle strip of electrodes from MEA

X_middle = pd.read_hdf('trial_data/achat_02_min_middle.h5')
# first try with just the middle strip
impute(X_middle)
X_middle_filtered = select_features(X_middle, y)

## Evaluation Method 1: split training and test data
X_middle_train, X_middle_test, y_train, y_test = train_test_split(X_middle_filtered, y, test_size=.4)

# TREE
tree_middle = tree.DecisionTreeClassifier()
tree_middle.fit(X_middle_train, y_train)
print(classification_report(y_test, tree_middle.predict(X_middle_test)))
print(confusion_matrix(y_test, tree_middle.predict(X_middle_test)))

# RANDOM FOREST
rf = RandomForestClassifier(n_estimators = 200, max_depth=3, random_state=0)
rf.fit(X_middle_train, y_train)
print(classification_report(y_test, rf.predict(X_middle_test)))
print(confusion_matrix(y_test, rf.predict(X_middle_test)))

## Evaluation Method 2: k-fold cross validation
# TREE
print('\n TREE \n Mean 3-fold cross-validation score = '+str(np.mean(cross_val_score(tree_middle, X_middle, y, cv=3))))

# RANDOM FOREST
print('\n RANDOM FOREST \n Mean 3-fold cross-validation score = '+str(np.mean(cross_val_score(rf, X_middle, y, cv=3))))

### Feature set 2: All electrodes from MEA

X = pd.read_hdf('trial_data/achat_02_min.h5')
# next try with the entire electrode array
impute(X)
X_filtered = select_features(X, y)

## Evaluation Method 1: split training and test data
X_train, X_test, y_train, y_test = train_test_split(X_filtered, y, test_size=.4)

# TREE
tree_full = tree.DecisionTreeClassifier()
tree_full.fit(X_train, y_train)
print('\n TREE \n')
print(classification_report(y_test, tree_full.predict(X_test)))
print(confusion_matrix(y_test, tree_full.predict(X_test)))

# RANDOM FOREST
print('\n RANDOM FOREST \n')
rf = RandomForestClassifier(n_estimators = 200, max_depth=3, random_state=0)
rf.fit(X_train, y_train)
print(classification_report(y_test, rf.predict(X_test)))
print(confusion_matrix(y_test, rf.predict(X_test)))

## Evaluation Method 2: k-fold cross validation
# TREE
print('\n TREE \n Mean 3-fold cross-validation score = '+str(np.mean(cross_val_score(tree_full, X_filtered, y, cv=3))))

# RANDOM FOREST
print('\n RANDOM FOREST \n Mean 3-fold cross-validation score = '+str(np.mean(cross_val_score(rf, X_filtered, y, cv=3))))

We observe that using the set of features generated from all of the electrodes provides slightly higher classification accuracy than only using the middle strip of electrodes. This confirms to me that we should be using all of the data when fitting a model to the whole dataset.  
Unsurprisingly, random forests perform better than individual trees. 3-fold cross-validation provides a less biased estimate of out-of-sample model accuracy. Additionally, we observe from the confusion matrices that each of the target variables share similar classification accuracies hence overall accuracy is not heavily biased.
