==> notebooks/9.2-kv-final_data_splits.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

# df = pd.read_hdf('../data/features/ach-at-hex_6000_eff_combined.h5')
y = pd.read_hdf('../data/processed/y_4_class_6000.h5')
subject = pd.read_hdf('../data/processed/subject_6000.h5')

mean = pd.read_hdf('../data/features/6000/ach-at-hex_6000_mean_eff.h5')
mean = mean.reset_index(drop=True)

df = df.join(mean)

y = y.reset_index(drop=True)
subject = subject.reset_index(drop=True)

subject.to_hdf('../data/FINAL/subject_base_v_ach_v_drug2.h5', key='data', complevel=9)

ach_v_at = y.copy()
ach_v_at = ach_v_at[ach_v_at>=1]
ach_v_at = ach_v_at[ach_v_at<=2]
ach_v_at[ach_v_at==1] = 'ach'
ach_v_at[ach_v_at==2] = 'at'
X_ach_v_at = df.iloc[ach_v_at.index,:]
subject_ach_v_at = subject[ach_v_at.index]



ach_v_at.to_hdf('../data/FINAL/y_ach_v_at.h5', key='data', complevel=9)
subject_ach_v_at.to_hdf('../data/FINAL/subject_ach_v_at.h5', key='data', complevel=9)
X_ach_v_at.to_hdf('../data/FINAL/X_ach_v_at.h5', key='data', complevel=9)

y_all = y.copy()
y_all[y==0] = 'baseline'
y_all[y==1] = 'ach'
y_all[y==2] = 'at'
y_all[y==3] = 'hex'
y_all.to_hdf('../data/FINAL/y_all.h5', key='data', complevel=9)

# MULTICLASS: Baseline vs. ACh vs. ATorHex
base_ach_drug2 = y.copy()
base_ach_drug2[y==0] = 'baseline'
base_ach_drug2[y==1] = 'ach'
base_ach_drug2[y==2] = 'drug2'
base_ach_drug2[y==3] = 'drug2'
# base_ach_drug2.to_hdf('../data/FINAL/y_base_v_ach_v_drug2.h5', key='data', complevel=9)

# MULTICLASS: Baseline vs. ACh vs. Hex
base_ach_hex = y.copy()
base_ach_hex = base_ach_hex[y!=2]
base_ach_hex[y==0] = 'baseline'
base_ach_hex[y==1] = 'ach'
base_ach_hex[y==3] = 'hex'
base_ach_hex.to_hdf('../data/FINAL/y_base_v_ach_v_hex.h5', key='data', complevel=9)
# X_base_ach_hex = df.iloc[base_ach_hex.index,:]
# subject_base_ach_hex = subject[base_ach_hex.index]

# X_base_ach_hex.to_hdf('../data/FINAL/X_base_v_ach_v_hex.h5', key='data', complevel=9)
# subject_base_ach_hex.to_hdf('../data/FINAL/subject_base_v_ach_v_hex.h5', key='data', complevel=9)

# MULTICLASS: Baseline vs. Ach vs. AT/Hex
multiclass = y.copy()
multiclass = multiclass[y!=3]
multiclass[y==0] = 'baseline'
multiclass[y==1] = 'ach'
multiclass[y==2] = 'at'
X_multiclass = df.iloc[multiclass.index,:]
subject_multiclass = subject[multiclass.index]
regex = '^1[8-9]__|^2[0-5]__'
X_multiclass_filtered = X_multiclass.filter(regex=regex, axis=1)

multiclass.to_hdf('../data/FINAL/y_base_v_ach_v_at.h5', key='data', complevel=9)
subject_multiclass.to_hdf('../data/FINAL/subject_base_v_ach_v_at.h5', key='data', complevel=9)
X_multiclass_filtered.to_hdf('../data/FINAL/X_base_v_ach_v_at_TOPLEFT.h5', key='data', complevel=9)

# baseline vs. rest
base_v_drug = y.copy()
base_v_drug[base_v_drug>=1] = 'drug'
base_v_drug[base_v_drug==0] = 'baseline'
base_v_drug.name = 'drug_applied'
base_v_drug.shape, base_v_drug.head(), base_v_drug.tail()

# ach vs. at or hex
first_v_second = y.copy()
first_v_second = first_v_second[first_v_second!=0]
first_v_second[first_v_second>1] = 'second'
first_v_second[first_v_second==1] = 'first'
first_v_second.name = 'drug_applied'
df_first_v_second = df.iloc[first_v_second.index,:]
subject_first_v_second = subject[first_v_second.index]
first_v_second.shape, first_v_second.head(), first_v_second.tail()

# at vs. hex
at_v_hex = y.copy()
at_v_hex = at_v_hex[at_v_hex>1]
at_v_hex[at_v_hex==2] = 'at'
at_v_hex[at_v_hex==3] = 'hex'
at_v_hex.name = 'drug_applied'
df_at_v_hex = df.iloc[at_v_hex.index,:]
subject_at_v_hex = subject[at_v_hex.index]
at_v_hex.shape, at_v_hex.head(), at_v_hex.tail()



base_v_ach = y.copy()
base_v_ach = base_v_ach[base_v_ach<2]
base_v_ach[base_v_ach==0] = 'baseline'
base_v_ach[base_v_ach==1] = 'ach'
base_v_ach.name = 'drug_applied'
df_base_v_ach = df.iloc[base_v_ach.index,:]
subject_base_v_ach = subject[base_v_ach.index]
base_v_ach.shape, base_v_ach.head(), base_v_ach.tail()

base_v_ach.to_hdf('../data/FINAL/y_base_v_ach.h5', key='data', complevel=9)
subject_base_v_ach.to_hdf('../data/FINAL/subject_base_v_ach.h5', key='data', complevel=9)
df_base_v_ach.to_hdf('../data/FINAL/X_base_v_ach.h5', key='data', complevel=9)

base_v_drug.to_hdf('../data/FINAL/y_base_v_drug.h5', key='data', complevel=9)
subject.to_hdf('../data/FINAL/subject_base_v_drug.h5', key='data', complevel=9)
df.to_hdf('../data/FINAL/X_base_v_drug.h5', key='data', complevel=9)

first_v_second.to_hdf('../data/FINAL/y_first_v_second.h5', key='data', complevel=9)
subject_first_v_second.to_hdf('../data/FINAL/subject_first_v_second.h5', key='data', complevel=9)
df_first_v_second.to_hdf('../data/FINAL/X_first_v_second.h5', key='data', complevel=9)

at_v_hex.to_hdf('../data/FINAL/y_at_v_hex.h5', key='data', complevel=9)
subject_at_v_hex.to_hdf('../data/FINAL/subject_at_v_hex.h5', key='data', complevel=9)
df_at_v_hex.to_hdf('../data/FINAL/X_at_v_hex.h5', key='data', complevel=9)


# baseline vs. at
base_v_at = y.copy()
base_v_at = base_v_at[(base_v_at==0) | (base_v_at==2)]
base_v_at[base_v_at==0] = 'baseline'
base_v_at[base_v_at==2] = 'at'
df_base_v_at = df.iloc[base_v_at.index,:]
subject_base_v_at = subject[base_v_at.index]
base_v_at.shape, base_v_at.head(), base_v_at.tail()


# baseline vs. hex
base_v_hex = y.copy()
base_v_hex = base_v_hex[(base_v_hex==0) | (base_v_hex==3)]
base_v_hex[base_v_hex==0] = 'baseline'
base_v_hex[base_v_hex==3] = 'hex'
base_v_hex.name = 'drug_applied'
df_base_v_hex = df.iloc[base_v_hex.index,:]
subject_base_v_hex = subject[base_v_hex.index]
base_v_hex.shape, base_v_hex.head(), base_v_hex.tail()

# ach vs. hex
ach_v_hex = y.copy()
ach_v_hex = ach_v_hex[(ach_v_hex==1) | (ach_v_hex==3)]
ach_v_hex[ach_v_hex==1] = 'ach'
ach_v_hex[ach_v_hex==3] = 'hex'
ach_v_hex.name = 'drug_applied'
df_ach_v_hex = df.iloc[ach_v_hex.index,:]
subject_ach_v_hex = subject[ach_v_hex.index]
ach_v_hex.shape, ach_v_hex.head(), ach_v_hex.tail()

ach_v_hex.to_hdf('../data/FINAL/y_ach_v_hex.h5', key='data', complevel=9)
subject_ach_v_hex.to_hdf('../data/FINAL/subject_ach_v_hex.h5', key='data', complevel=9)
df_ach_v_hex.to_hdf('../data/FINAL/X_ach_v_hex.h5', key='data', complevel=9)

base_v_hex.to_hdf('../data/FINAL/y_base_v_hex.h5', key='data', complevel=9)
subject_base_v_hex.to_hdf('../data/FINAL/subject_base_v_hex.h5', key='data', complevel=9)
df_base_v_hex.to_hdf('../data/FINAL/X_base_v_hex.h5', key='data', complevel=9)

base_v_at.to_hdf('../data/FINAL/y_base_v_at.h5', key='data', complevel=9)
subject_base_v_at.to_hdf('../data/FINAL/subject_base_v_at.h5', key='data', complevel=9)
df_base_v_at.to_hdf('../data/FINAL/X_base_v_at.h5', key='data', complevel=9)
