==> notebooks/9.4-kv-topleft-randomforest.ipynb <==
import sys
sys.path.insert(0, "..")
%load_ext autoreload
%autoreload 2

from joblib import load
import pandas as pd
from sklearn.model_selection import GridSearchCV, GroupKFold, cross_val_score
from sklearn.ensemble import RandomForestClassifier
from src.features.modified_feature_selector import FeatureSelector
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

X = pd.read_hdf('../data/FINAL/X_base_v_ach_v_at_TOPLEFT.h5')
y = pd.read_hdf('../data/FINAL/y_base_v_ach_v_at.h5')
subject = pd.read_hdf('../data/FINAL/subject_base_v_ach_v_at.h5')

fs = FeatureSelector(multiclass=True, ml_task='classification', n_significant=3)
fs.fit(X,y)
X_filt = fs.transform(X)

rf = RandomForestClassifier(n_estimators=100, max_depth=None)
gkf = GroupKFold(n_splits = len(subject.unique()))
gkf = list(gkf.split(X, y, subject))

cross_val_score(rf, X_filt, y, cv=gkf, n_jobs=8)

np.mean([0.33333333, 0.31111111, 0.47777778, 0.52222222, 0.42222222,
       0.68333333, 0.81666667, 0.63333333, 0.68333333, 0.53333333,
       0.71666667])
